requirements.txt
asgiref==3.2.10
certifi==2020.11.8
chardet==3.0.4
Django==3.1.2
django-oauth-toolkit==1.3.3
djangorestframework==3.12.2
idna==2.10
mccabe==0.6.1
oauthlib==3.1.0
Pillow==8.0.0
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
pylama==7.7.1
pytz==2020.4
requests==2.25.0
snowballstemmer==2.0.0
sqlparse==0.4.1
urllib3==1.26.2
